requirements.txt
django
djangorestframework
django-environ
django-cors-headers
django-filter
djangorestframework-simplejwt